pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools"]

[project]
name = "PataNgoma"
version = "1.0.0"
authors = [
  { name = "Fourtythree43", email = "shaqmwa@outlook.com" },
  { name = "Kemboiray", email = "kemboiray@gmail.com" },
  { name = "Fourtythree43", email = "shaqmwa@outlook.com" }
]
classifiers = [
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3"
]
dependencies = [
  "click",
  "deezer-python",
  "fuzzywuzzy",
  "imgcat",
  "inquirerpy",
  "mediafile",
  "musicbrainzngs",
  "pillow",
  "pytest",
  "python-dotenv",
  "python3-discogs-client",
  "pyyaml",
  "spotipy",
  "rich",
  "rgbprint",
  "pyfiglet",
  "toml",
  "cachetools"
]
description = """
A command-line tool to enhance your music library's metadata from multiple APIs
"""
readme = "README.md"
requires-python = ">=3.7"

[project.urls]
"Homepage" = "https://github.com/FourtyThree43/PataNgoma-AudioTagger-tool"
"Bug Tracker" = "https://github.com/FourtyThree43/PataNgoma-AudioTagger-tool/issues"
